requirements.txt
psutil==5.9.8
pygetwindow==0.0.9
pyautogui==0.9.54
pystray==0.19.5
pillow==10.2.0
plyer==2.1.0
schedule==1.2.1
pygame==2.5.2 